pyproject.toml
[project]
name = "feishu2everywhere"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # "selenium",
    # "webdriver-manager",
    # "requests",
    # "beautifulsoup4",

    # with version
    "selenium==4.11.0",
    "webdriver-manager==3.8.5",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "pyyaml==6.0.1",
    "keyboard==0.13.3",
]
